refactor: create helper function for testing weighted lists
