+++ kotlin-utils/src/main/kotlin/kotlinutils/list/ListAliases.kt
@@ -14,3 +14,8 @@ typealias IntList = List<Int>
  * List of Strings
  */
 typealias StringList = List<String>
+
+/**
+ * List of values paired with weights
+ */
+typealias WeightedList<T> = List<Pair<T, Float>>
